Adding C#, ExportChecker and other
ExportChecker is a class that checks passed objects for nulls these object should be exported node values which will be null if not set, if a null is found then an error is logged in a log file and Godot debugger.

Directory changes:
Do not delete .gdignore it's to ignore it from being built into the game.

- exports is to keep all exports organized in the same directory.
- game_data is there for logs files and other dynamically created game files, It allows some game features to be tested. DO NOT upload any files into this directory.