Added setting menu scene and crucials
Added the setting menu scene and all the scripts required for it to function.

Some important ones:
UserSettingControllerNode - Saves settings and passes it to a coupler in order to save or cancel the settings.
BasicDropdown - Special dropdown node created because the default OptionButton was very limited to customizing the list, changes visibility of an exported list and changes text to the selected text in the list.
AdvancedScrollbar - Special scrollbar made so the scrolling was possible and that the inside scrollbars would scroll.