Added Json attributes and converters
FallbackJsonConverter attribute - Provides a way to give a fallback value if the For now can only be used with converters which have a single variable in their constructor, the variable being the fallback value.

DefaultEnumConverter - Used in the fallback attribute on Enums it has to be constructed with the fallback value. Places the fallback value when the conversion for string to Enum value fails.

NullEnumConverter - Used on Enums, places a null if conversion of enum from string fails Can be used in simple JsonConverter attribute and requires that the enum property to be nullable by for example placing ? at the end.