Added GameLoaderMenu, functionalities and other
This update added a lot of functionalities to the game. Now you can actually load the game and create it!

Some of those:

GameLoaderMenu is a menu that presents a list of saves, which you can either load or delete (both require confirmation).

GameSaveHelper is a class that helps with saves, like creating or deleteing them, getting the most recent save etc.

ConfirmationDialog and derivatives, It's a special class that allows other classes (like controllers) to plug in to their Decision event.

GameSaveItem, is a special class that provides specific events that allow the controller classes to manage it more cleanly.

GameSaveManager allows for saving and loading the saves.